requirements.txt
pytest==8.0.0
pytest-cov==4.1.0
requests==2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
pyyaml==6.0.1
